worker/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
redis==5.2.0
python-dotenv==1.0.1
aiofiles==24.1.0
azure-storage-blob==12.23.1

